refactor(git): split mod.rs into remote.rs and worktree.rs
Cleanup-pass finding #5 from the code-quality scan. No behavior changes.

Before: src/git/mod.rs was 1,797 lines hosting three distinct concerns.
After:
  - src/git/mod.rs (35 lines) — module declarations, re-exports, and the
    shared pub(crate) open_repo_at helper used by sibling submodules.
  - src/git/remote.rs (169 lines) — clone_repo, parse_owner_from_remote_url,
    get_remote_owner, and their parse tests.
  - src/git/worktree.rs (1,624 lines) — GitWorktree lifecycle methods,
    branch ops, template path expansion, and all the worktree tests.

Public API (crate::git::{GitWorktree, WorktreeEntry, clone_repo,
get_remote_owner}) is preserved, so downstream consumers in cli/,
session/, and server/ are unchanged.
